refactor: 抽离getAggregationAndCalcFuncByQuery到utils
